% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DrugZR.R
\name{DrugZR}
\alias{DrugZR}
\title{DrugZR
Identify drug-gene interactions in paired sample genomic perturbation screens}
\usage{
DrugZR(
  se,
  control,
  treatment,
  min_observations = 1,
  pseudocount = 5,
  seTrue = TRUE
)
}
\description{
Version 1.0.0
}
\details{
Last Modified Oct 14, 2024


inputs: SummarizeExperiment (se) / DrugZ countTable

input parameters:

pseduocount:Constant value added to all reads (default 5) - prevents log(0) problems

min_observations: An integer value to act as a threshold for the minimum number observations to be included in the analysis (default=1)

seTrue: whether the input is SummarizedExperiment (default TRUE)
}
